update uvwasi to version 0.0.8 (#132)
* update uvwasi to version 0.0.8

This release focuses on improving the robustness of the path
resolution and sandboxing, including support for relative
preopen paths.

Also remove the ${} around the uvwasi version, which was a
mistake. The result was that the latest master was always
installed.

* test: update wasi suite to test relative paths

This commit makes sure that both absolute and relative paths get
some test coverage.

* test: use uvwasi on alpine

Run the WASI tests on Alpine using uvwasi. This requires
installing git as well.